requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
pydantic>=1.10.13
google-adk>=1.18.0